refactor asset building and all of the other 11ty config stuff
