+++ _includes/scripts.html
@@ -1,5 +1,5 @@
 <!-- Pull in USWDS and custom js -->
-<script async src="{{ assetPaths['app.js'] }}"></script>
+<script async src="{% link "/assets/js/app.js" %}"></script>
 
 {% if site.dap.agency %}
   <!-- Digital Analytics Program roll-up, see https://analytics.usa.gov for data -->
