+++ config/shortcodes/page.js
@@ -1,0 +1,4 @@
+const path = require("path");
+
+// Given a page reference, turn it into a link on the published site.
+module.exports = (link) => path.join(process.env.BASEURL ?? "/", link);
