+++ package.json
@@ -3,22 +3,13 @@
   "version": "3.0.0",
   "description": "The TTS Handbook",
   "scripts": {
-    "build": "npm run assets:build && npx @11ty/eleventy",
-    "assets:autoprefix": "postcss _site/assets/styles/*.css -r --use autoprefixer",
-    "assets:build": "node ./config/buildAssets && npm run assets:autoprefix",
-    "assets:clean": "rimraf _site/assets",
-    "assets:refresh": "npm run assets:clean && npm run assets:build && npx @11ty/eleventy",
-    "assets:watch": "chokidar 'js/**' 'styles/**' -c 'npm run assets:refresh' --polling",
-    "clean": "rimraf _site",
-    "dev": "npm run clean && npm-run-all -p dev:assets dev:serve",
-    "dev:clean": "npm run clean && npm run dev",
-    "dev:assets": "npm run assets:refresh && npm run assets:watch",
-    "dev:debug": "DEBUG=* npx @11ty/eleventy --serve",
-    "dev:serve": "npx @11ty/eleventy --serve --incremental",
+    "build": "eleventy",
+    "clean": "rm -rf _site",
+    "debug": "DEBUG=* eleventy --serve --incremental",
     "federalist": "npm run build",
     "lint": "prettier -w .",
-    "serve": "npx @11ty/eleventy --serve --incremental",
-    "start": "npx @11ty/eleventy --serve --incremental",
+    "serve": "npm run start",
+    "start": "eleventy --serve --incremental",
     "test": "npm run test:prefixed-links && npm run test:internal-links && npm run test:html-validation",
     "test:html-validation": "html-validate _site/**/*.html",
     "test:internal-links": "node check-links.js",
@@ -26,30 +17,31 @@
   },
   "author": "",
   "license": "CC0-1.0",
-  "devDependencies": {
+  "dependencies": {
     "@11ty/eleventy": "^2.0.1",
     "@11ty/eleventy-img": "^3.1.0",
     "autoprefixer": "^10.4.2",
-    "cheerio": "^1.0.0-rc.12",
-    "chokidar-cli": "^3.0.0",
     "eleventy-plugin-svg-sprite": "^2.4.0",
     "esbuild": "^0.17.19",
     "esbuild-sass-plugin": "^2.9.0",
-    "html-validate": "^7.18.1",
     "js-yaml": "^4.1.0",
     "markdown-it": "^13.0.1",
     "markdown-it-anchor": "^8.6.7",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.24",
+    "postcss": "^8.4.23",
     "postcss-cli": "^10.0.0",
-    "prettier": "^2.8.8",
-    "rimraf": "^4.4.1"
-  },
-  "dependencies": {
     "@uswds/uswds": "^3.2.0",
     "glob": "^8.1.0",
     "jsdom": "^21.1.2",
     "markdown-it-attrs": "^4.1.6"
+  },
+  "devDependencies": {
+    "cheerio": "^1.0.0-rc.12",
+    "eslint": "^8.41.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "html-validate": "^7.16.0",
+    "prettier": "^2.8.8"
   },
   "prettier": {
     "proseWrap": "always"
